Add dialog for updating portfolio holdings on the portfolio page
