+++ apps/backend/src/models/userAccount.entity.ts
@@ -40,15 +40,23 @@ export class UserAccount extends BaseEntity {
   // })
   // role: UserRole;
 
-  // TODO: Add Date Created column
+  // TODO: Add Date Joined column
+  // @Column({ name: 'date_joined', type: 'timestamptz', default: () => 'CURRENT_TIMESTAMP', nullable: false })
+  // dateJoined: Date;
 
   // TODO: Add Last Login column
+  // @Column({ name: 'last_login', type: 'timestamptz', default: () => 'CURRENT_TIMESTAMP', nullable: false })
+  // lastLogin: Date;
 
   // TODO: add an about column
 
   // TODO: add an active column
+  // @Column({ type: 'boolean', nullable: false, default: true })
+  // isActive: boolean;
 
   // TODO: Add verified column
+  // @Column({ type: 'boolean', nullable: false, default: false })
+  // emailVerified: boolean;
 
   @Column({
     name: 'spirit_animal',
