+++ apps/frontend/src/app/core/header/header.component.html
@@ -5,7 +5,7 @@
     </button>
     <a routerLink="/">
       <span id="header-wrapper"
-        ><img src="../../../assets/images/logos/rms_primary_logomark.png" id="header-logo"
+        ><img src="../../../assets/images/logos/rms_primary_logomark_beta.png" id="header-logo"
       /></span>
     </a>
