+++ apps/frontend/src/app/modules/login/components/auth-form-container/auth-form-container.component.html
@@ -1,7 +1,7 @@
 <div class="authContainer">
   <mat-card>
     <div class="login-header">
-      <img src="../../../../assets/images/logos/rms_primary_logomark.png" />
+      <img src="../../../../assets/images/logos/rms_primary_logomark_beta.png" />
     </div>
     <ng-content>
       <!-- Child form components will be displayed here -->
