+++ apps/frontend/src/app/modules/portfolio/components/update-portfolio-holdings-dialog/update-portfolio-holdings-dialog.component.html
@@ -1,1 +1,15 @@
-<p>update-portfolio-holdings-dialog works!</p>
+<div class="update-portfolio-holdings-dialog-container">
+  <h2 mat-dialog-title>Update Holdings</h2>
+  <form [formGroup]="portfolioForm" (ngSubmit)="onSubmit()">
+    <mat-dialog-content class="mat-typography">
+      <label><strong>Add a stock:</strong></label>
+      <app-stock-search [placeholder]="'Ticker/Name'" [redirect]="false"></app-stock-search>
+
+      <app-portfolio-holdings-table [portfolioStocks]="portfolioStocks" #portfolioTable></app-portfolio-holdings-table>
+    </mat-dialog-content>
+    <mat-dialog-actions align="end">
+      <button mat-raised-button color="warn" mat-dialog-close>Cancel</button>
+      <button type="submit" mat-raised-button [disabled]="portfolioForm.invalid" color="primary">Update</button>
+    </mat-dialog-actions>
+  </form>
+</div>
