+++ apps/frontend/src/app/modules/portfolio/components/update-portfolio-holdings-dialog/update-portfolio-holdings-dialog.component.spec.ts
@@ -1,6 +1,10 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { UpdatePortfolioHoldingsDialogComponent } from './update-portfolio-holdings-dialog.component';
+import { HttpClientTestingModule } from '@angular/common/http/testing';
+import { PortfolioModule } from '../../portfolio.module';
+import { MAT_DIALOG_DATA, MatDialogRef } from '@angular/material/dialog';
+import { RouterTestingModule } from '@angular/router/testing';
 
 describe('UpdatePortfolioHoldingsDialogComponent', () => {
   let component: UpdatePortfolioHoldingsDialogComponent;
@@ -8,9 +12,13 @@ describe('UpdatePortfolioHoldingsDialogComponent', () => {
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      declarations: [ UpdatePortfolioHoldingsDialogComponent ]
-    })
-    .compileComponents();
+      declarations: [UpdatePortfolioHoldingsDialogComponent],
+      imports: [HttpClientTestingModule, RouterTestingModule, PortfolioModule],
+      providers: [
+        { provide: MAT_DIALOG_DATA, useValue: { portfolioStocks: [] } },
+        { provide: MatDialogRef, useValue: {} },
+      ],
+    }).compileComponents();
   });
 
   beforeEach(() => {
